kematian role

+++ application/views/data/ibu_hamil/kematian/index.php
@@ -9,9 +9,11 @@
                <h6 class="m-0 font-weight-bold text-primary"><?= $title ?></h6>
             </div>
             <div class="col-6">
-               <button type="button" class="btn btn-primary float-right" data-toggle="modal" data-target="#addModal">
-                  <i class="fas fa-plus"></i> Tambah
-               </button>
+               <?php if ($role == 7) : ?>
+                  <button type="button" class="btn btn-primary float-right" data-toggle="modal" data-target="#addModal">
+                     <i class="fas fa-plus"></i> Tambah
+                  </button>
+               <?php endif; ?>
             </div>
          </div>
       </div>
@@ -30,7 +32,9 @@
                      <th>Jenis</th>
                      <th>Meninggal di</th>
                      <th>Keterangan</th>
-                     <th>Aksi</th>
+                     <?php if ($role == 7) : ?>
+                        <th>Aksi</th>
+                     <?php endif; ?>
                   </tr>
                </thead>
                <tbody>
@@ -44,10 +48,12 @@
                         <td><?= $field['jenis'] ?></td>
                         <td><?= $field['meninggal_di'] ?></td>
                         <td><?= $field['keterangan'] ?></td>
-                        <td>
-                           <button type="button" class="btn btn-block btn-primary" data-toggle="modal" data-target="#editModal" onclick="getData(<?= htmlspecialchars(json_encode($field), ENT_QUOTES, 'UTF-8') ?>)">Edit</button>
-                           <button type="button" class="btn btn-block btn-danger" data-toggle="modal" data-target="#deleteModal" onclick="deleteData(<?= htmlspecialchars(json_encode($field), ENT_QUOTES, 'UTF-8') ?>)">Hapus</button>
-                        </td>
+                        <?php if ($role == 7) : ?>
+                           <td>
+                              <button type="button" class="btn btn-block btn-primary" data-toggle="modal" data-target="#editModal" onclick="getData(<?= htmlspecialchars(json_encode($field), ENT_QUOTES, 'UTF-8') ?>)">Edit</button>
+                              <button type="button" class="btn btn-block btn-danger" data-toggle="modal" data-target="#deleteModal" onclick="deleteData(<?= htmlspecialchars(json_encode($field), ENT_QUOTES, 'UTF-8') ?>)">Hapus</button>
+                           </td>
+                        <?php endif; ?>
                      </tr>
                   <?php endforeach; ?>
                </tbody>
